--- les_2_new/ex_2.py
@@ -1,18 +1,0 @@
-"""
-Сформировать из введенного числа обратное по порядку входящих в него цифр и вывести на экран. 
-Например, если введено число 3486, то надо вывести число 6843. (Сделать без использования строк)
-"""
-
-
-def backward_rec(n: int) -> None:
-    if n // 10 == 0:
-        print(n % 10, end='')
-    else:
-        print(n % 10, end='')
-        backward(n // 10)
-
-
-
-if __name__=='__main__':
-    n = int(input("Сформируем из введенного числа обратное\n>>> "))
-    backward_rec(n)
