--- les_2_new/ex_3.py
@@ -1,18 +1,0 @@
-"""
-    Найти сумму n элементов следующего ряда чисел: 1 -0.5 0.25 -0.125 ...
-    Количество элементов (n) вводится с клавиатуры.
-"""
-
-
-def row_sum(n: int) -> None:
-    item = 1
-    result = 0
-    for _ in range(n):
-        result += item
-        item /= -2
-    print(result)
-
-if __name__ == '__main__':
-    n = int(input('Найдём сумму n элементов следующего ряда чисел: 1 -0.5 0.25 -0.125 ...\n'
-                  'Введите n >>> '))
-    row_sum(n)
