--- les_2_new/ex_4.py
@@ -1,18 +1,0 @@
-"""
-Напишите программу, доказывающую или проверяющую, что для множества натуральных чисел
-выполняется равенство: 1+2+...+n = n(n+1)/2, где n - любое натуральное число.
-"""
-
-def equation(n: int):
-    a = 0
-    for i in range(n):
-        a += i + 1
-    b = n * (n + 1) / 2
-    print(f'1+2+...+n = {a}\nn(n+1)/2 = {int(b)}')
-
-
-if __name__ == '__main__':
-    n = int(input('Введите число n\n>>> '))
-    equation(n)
-
-
