+++ lesson_2/ex_2.py
@@ -1,30 +1,18 @@
 """
-    Посчитать четные и нечетные цифры введенного натурального числа.
-    Например, если введено число 34560, то у него 3 четные цифры (4, 6 и 0) и 2 нечетные (3 и 5).
+Сформировать из введенного числа обратное по порядку входящих в него цифр и вывести на экран. 
+Например, если введено число 3486, то надо вывести число 6843. (Сделать без использования строк)
 """
 
 
-def count_even_odd_num():
-
-    number = input('Считаем четные и нечетные цифры числа.\n'
-                    'Ведите натуральное число:\n>>> ')
-
-    if number.isdigit():
-        numbers_list = list(number)
+def backward_rec(n: int) -> None:
+    if n // 10 == 0:
+        print(n % 10, end='')
     else:
-        print('Ошибка ввода')
-        return
-
-    even_count = 0
-    odd_count = 0
-    for el in range(len(numbers_list)):
-        if int(el) & 1:
-            odd_count += 1
-        else:
-            even_count += 1
-
-    print(f'В числе {number}\nНечетных цифр: {odd_count}\nЧетных цифр: {even_count}')
+        print(n % 10, end='')
+        backward(n // 10)
 
 
-if __name__ == '__main__':
-    count_even_odd_num()
+
+if __name__=='__main__':
+    n = int(input("Сформируем из введенного числа обратное\n>>> "))
+    backward_rec(n)
