+++ lesson_2/ex_3.py
@@ -1,28 +1,18 @@
 """
-    Сформировать из введенного числа обратное по порядку входящих в него цифр и вывести на экран.
-    Например, если введено число 3486, то надо вывести число 6843.
+    Найти сумму n элементов следующего ряда чисел: 1 -0.5 0.25 -0.125 ...
+    Количество элементов (n) вводится с клавиатуры.
 """
 
 
-def reversed_num():
-
-    number = input('Сформируем из введенного числа обратное\n'
-                    'Ведите число:\n>>> ')
-
-    if number.isdigit():
-        print(number[::-1])
-
-    # Другой вариант
-    # for el in range(0, len(number), -1):
-    #     result = ''.join(el)
-    #     print(result)
-    #     return
-
-    else:
-        print('Ошибка ввода')
-        return
-
+def row_sum(n: int) -> None:
+    item = 1
+    result = 0
+    for _ in range(n):
+        result += item
+        item /= -2
+    print(result)
 
 if __name__ == '__main__':
-    reversed_num()
-
+    n = int(input('Найдём сумму n элементов следующего ряда чисел: 1 -0.5 0.25 -0.125 ...\n'
+                  'Введите n >>> '))
+    row_sum(n)
