+++ lesson_2/ex_4.py
@@ -1,35 +1,18 @@
 """
-    Найти сумму n элементов следующего ряда чисел: 1 -0.5 0.25 -0.125 ...
-    Количество элементов (n) вводится с клавиатуры.
+Напишите программу, доказывающую или проверяющую, что для множества натуральных чисел
+выполняется равенство: 1+2+...+n = n(n+1)/2, где n - любое натуральное число.
 """
 
-
-def validation(text):
-    try:
-        n = int(input(text))
-        if n < 0:
-            raise ValueError
-        return n
-    except ValueError:
-        print('Ошибка ввода')
-        return
-
-
-def row_sum():
-    n = validation('Найдём сумму n элементов следующего ряда чисел: 1 -0.5 0.25 -0.125 ...\n'
-                      'Введите колличество элементов n\n>>> ')
-    if n == None: return
-
-    result = 0
-    item = 1
-    row = ''
+def equation(n: int):
+    a = 0
     for i in range(n):
-        result += item
-        row += ' ' + str(item)
-        item /= -2
-    print(f'Сумма ряда {row} равна {result}')
-
+        a += i + 1
+    b = n * (n + 1) / 2
+    print(f'1+2+...+n = {a}\nn(n+1)/2 = {int(b)}')
 
 
 if __name__ == '__main__':
-    row_sum()
+    n = int(input('Введите число n\n>>> '))
+    equation(n)
+
+
