+++ lesson_2/ex_5.py
@@ -1,17 +1,17 @@
 """
-Вывести на экран коды и символы таблицы ASCII, начиная с символа под номером 32 и
-заканчивая 127-м включительно. Вывод выполнить в табличной форме: по десять пар "код-символ" в каждой строке.
+    Посчитать, сколько раз встречается определенная цифра в введенной последовательности чисел.
+    Количество вводимых чисел и цифра, которую необходимо посчитать, задаются вводом с клавиатуры.
 """
 
-
-def ascii_table():
-    n = 32
-    while n < 128:
-        m = n + 10
-        [print(f'{i} : {chr(i)}', end='\t') for i in range(n, m) if i < 128]
-        print()
-        n += 10
+def digit_count(numbers: str, digit: str) -> None:
+    count = 0
+    for i in numbers:
+        if i == digit:
+            count += 1
+    print(count)
 
 
-if __name__ == '__main__':
-    ascii_table()
+if __name__=="__main__":
+    numbers = input("Введите последовательность чисел\n>>> ")
+    digit = input("Введите цифру для подсчета\n>>> ")
+    digit_count(numbers, digit)
