+++ lesson_2_old_9_ex/ex_5.py
@@ -1,0 +1,17 @@
+"""
+Вывести на экран коды и символы таблицы ASCII, начиная с символа под номером 32 и
+заканчивая 127-м включительно. Вывод выполнить в табличной форме: по десять пар "код-символ" в каждой строке.
+"""
+
+
+def ascii_table():
+    n = 32
+    while n < 128:
+        m = n + 10
+        [print(f'{i} : {chr(i)}', end='\t') for i in range(n, m) if i < 128]
+        print()
+        n += 10
+
+
+if __name__ == '__main__':
+    ascii_table()
