+++ lesson_2_old_9_ex/ex_8.py
